feat(hunt): allow publishing a Piglet with no prize (optional LNURL)
Step 2 was labelled "Make the prize (optional)" but the only path to a
publishable stage was handleValidate, which hard-returned when the LNURL
field was empty — so the prize was never actually optional.

Add a `noPrize` Stage variant and a "Skip prize — publish without reward"
affordance on step 2. It flips the wizard into a publishable state without
a withdraw link; the public toggle, Publish button and step-5 helper all
treat `noPrize` like `validated`. Typing/scanning an LNURL supersedes the
skip choice (resets to idle so Validate re-engages), and a "Add a prize
instead" escape hatch returns from the no-reward card.

With no LNURL, listingIsLp is false in handleSave, so the published kind
37516 event carries no L/l label and no amount/wait/uses tags — a plain
NIP-GC cache. This invariant is already covered by the buildCacheListing
unit tests (lnurlw:'' + isLpPiggy:false omits amount/label).

To keep the over-cap HuntCreateScreen shrinking (CLAUDE.md), lift the
wizard's pure presentational chrome (StepHeader, StepProgressBar,
StepNavRow, LevelPicker, OptionPicker, NfcSupportedTagsCard) into
src/components/HuntCreateWizardChrome.tsx. Screen drops 2386 → 2211 lines.
